+++ composeApp/src/commonMain/kotlin/com/altmann/choresmanager/models/chores/grocery/Item.kt
@@ -1,7 +1,7 @@
 package com.altmann.choresmanager.models.chores.grocery
 
 class Item(
-    val itemId : Int,
+    val itemId: String,
     val name: String,
     var quantity: Int
 )